# docs/openapi.yaml
openapi: 3.0.3
info:
  title: NanoPay API
  version: "2.0.0"
  description: |
    API Backend NanoPay — commande centrale unique reliant Client, Commerçant, Banque et Admin,
    avec trois modes de financement : **Épargne**, **Coffre** (prélèvement sur salaire) et **Crédit**.

    Conforme au « Cahier de construction BACKEND ».

    ## Contrat de connexion
    | Élément | Contrat |
    |---|---|
    | Base URL | `/api` (alias historique `/api/v1`) |
    | Format | JSON UTF-8 (multipart pour les documents) |
    | Auth | `Authorization: Bearer <accessToken>` + refresh token (`/auth/refresh`, rotation) |
    | Succès | `{ "data": ..., "meta"?: ... }` |
    | Erreur | `{ "code", "message", "details"?, "requestId" }` (`X-Request-Id` aussi en en-tête) |
    | Pagination | query `page`, `limit` ; `meta` = `{ page, limit, total, nextPage }` |
    | Dates | ISO 8601 |
    | Montants | FCFA **entiers**, jamais de flottant |
    | Identifiants | UUID serveur ; `orderNumber` lisible (ex. `NP-2026-00001245`) pour l'UI |
    | Temps réel | pas de WebSocket : polling contrôlé de `GET /notifications` et des ressources |

    Les règles financières et de sécurité sont **toujours** appliquées côté backend ; le frontend
    ne fait que consommer ces contrats.

    ## Circuit
    Client → produit → commande (`CREEE`) → `select-financing` (`FINANCEMENT_EN_COURS`) →
    Épargne / Coffre / Crédit → `FINANCEE` → `PRETE_A_LIVRER` (+ QR) → scan commerçant →
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /marketplace/communes:
    get:
      tags: [Marketplace]
      summary: Communes d'une ville
      security: []
      parameters:
        - { name: cityId, in: query, required: true, schema: { type: string } }
      responses:
        "200": { description: "`data` = [{ id, cityId, name }]" }
  /stores:
    get:
      tags: [Marketplace]
      summary: Boutiques (commerçants validés uniquement)
      security: []
      parameters:
        - { name: cityId, in: query, schema: { type: string } }
        - { name: communeId, in: query, schema: { type: string } }
        - { $ref: "#/components/parameters/Page" }
        - { $ref: "#/components/parameters/Limit" }
      responses:
        "200": { description: Page de boutiques }
  /stores/{storeId}:
    get:
      tags: [Marketplace]
      summary: Détail d'une boutique et de ses produits publiés en stock
      security: []
      parameters:
        - { name: storeId, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Boutique }
        "404": { $ref: "#/components/responses/Error" }
  /products/{productId}:
    get:
      tags: [Marketplace]
      summary: Détail d'un produit (prix FCFA entier, définitif une fois publié)
      security: []
      parameters:
        - { name: productId, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Produit }
        "404": { $ref: "#/components/responses/Error" }
  # ------------------------------------------------------------------ ORDERS
  /orders:
    post:
      tags: [Orders]
      summary: Crée la commande d'un produit (CLIENT)
      description: Réserve une unité de stock. Le prix vient toujours de la base. Statut initial `CREEE`.
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [productId], properties: { productId: { type: string } } } } }
      responses:
        "201": { description: Commande créée avec `orderNumber` }
        "404": { $ref: "#/components/responses/Error" }
        "409": { description: Produit épuisé, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
    get:
      tags: [Orders]
      summary: Mes commandes (CLIENT)
      parameters:
        - { name: status, in: query, schema: { $ref: "#/components/schemas/OrderStatus" } }
        - { $ref: "#/components/parameters/Page" }
        - { $ref: "#/components/parameters/Limit" }
      responses:
        "200": { description: Page de commandes }
  /orders/{id}:
    get:
      tags: [Orders]
      summary: Détail d'une commande (propriétaire uniquement, sinon 404)
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        "200": { description: Commande avec épargne / coffre / demande de crédit / crédit }
        "404": { $ref: "#/components/responses/Error" }
  /orders/{id}/select-financing:
    post:
      tags: [Orders]
      summary: Choisit le mode de financement
      description: |
        `VAULT` et `CREDIT` exigent un profil crédit bancaire déjà `VALIDE` (Coffre : salarié uniquement).
        La commande passe à `FINANCEMENT_EN_COURS`.
      parameters: [{ $ref: "#/components/parameters/Id" }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [mode], properties: { mode: { type: string, enum: [SAVINGS, VAULT, CREDIT] } } } } }
      responses:
        "200": { description: Commande mise à jour }
        "403": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /orders/{id}/qr:
    get:
      tags: [Orders]
      summary: Jeton QR serveur de la commande prête à livrer
      description: |
        `data` = `{ qrToken, expiresAt }`. Le QR ne contient que ce jeton opaque. Un jeton expiré est
        régénéré (l'ancien est invalidé). 403 tant que le financement n'est pas validé.
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        "200": { description: Jeton QR }
        "403": { $ref: "#/components/responses/Error" }
  # ------------------------------------------------------------------ SAVINGS
...
  /merchant/store:
    get: { tags: [Merchant], summary: Ma boutique, responses: { "200": { description: Boutique } } }
    post:
      tags: [Merchant]
      summary: Crée ma boutique (compte VALIDE requis)
      requestBody: { $ref: "#/components/requestBodies/Store" }
      responses: { "201": { description: Boutique créée }, "403": { $ref: "#/components/responses/Error" } }
    put:
      tags: [Merchant]
      summary: Met à jour ma boutique
      requestBody: { $ref: "#/components/requestBodies/Store" }
      responses: { "200": { description: Boutique mise à jour } }
  /merchant/products:
    post:
      tags: [Merchant]
      summary: Publie un produit
      description: "`price` : FCFA entier, **définitif** après publication."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, price, stock]
              properties:
                title: { type: string }
                description: { type: string }
                price: { type: integer, minimum: 1 }
                stock: { type: integer, minimum: 0 }
                isPublished: { type: boolean }
      responses: { "201": { description: Produit créé } }
  /merchant/products/{id}:
    patch:
      tags: [Merchant]
      summary: Modifie titre, description, stock ou publication
      description: Toute tentative de modifier `price` est refusée (422).
      parameters: [{ $ref: "#/components/parameters/Id" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                description: { type: string }
                stock: { type: integer, minimum: 0 }
                isPublished: { type: boolean }
      responses: { "200": { description: Produit mis à jour }, "422": { $ref: "#/components/responses/Error" } }
    delete:
      tags: [Merchant]
      summary: Supprime un produit (suppression logique)
      description: Refusée (409) tant qu'une commande non terminée le référence.
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses: { "204": { description: Supprimé }, "409": { $ref: "#/components/responses/Error" } }
  /merchant/products/{id}/media:
    post:
      tags: [Merchant]
      summary: Ajoute photos (`photos`, 10 max) et spots publicitaires (`spots`, 5 max)
      parameters: [{ $ref: "#/components/parameters/Id" }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                photos: { type: array, items: { type: string, format: binary } }
                spots: { type: array, items: { type: string, format: binary } }
      responses: { "201": { description: Médias ajoutés } }
  /merchant/orders:
    get:
      tags: [Merchant]
      summary: Commandes de ma boutique
      parameters:
        - { name: status, in: query, schema: { $ref: "#/components/schemas/OrderStatus" } }
        - { $ref: "#/components/parameters/Page" }
        - { $ref: "#/components/parameters/Limit" }
      responses: { "200": { description: Page de commandes } }
  /merchant/settlements:
    get:
      tags: [Merchant]
      summary: Mes règlements
      parameters: [{ $ref: "#/components/parameters/Page" }, { $ref: "#/components/parameters/Limit" }]
      responses: { "200": { description: Page de règlements } }
  /merchant/qr/scan:
    post:
      tags: [Merchant]
      summary: Valide un QR côté serveur (aucun changement d'état)
      description: |
        404 jeton inconnu ou d'une autre boutique (indistinguables) ; 409 déjà utilisé ou commande non prête ;
        410 `QR_EXPIRED`.
      requestBody: { $ref: "#/components/requestBodies/QrToken" }
      responses:
        "200": { description: "`data` = { orderId, orderNumber, product, amount, financingMode, client, qrExpiresAt }" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "410": { $ref: "#/components/responses/Error" }
  /merchant/orders/{id}/withdrawal-confirm:
    post:
      tags: [Merchant]
      summary: Confirme le retrait en boutique
      description: Consomme le jeton (atomique), commande → `LIVREE`, règlement commerçant créé (`EN_ATTENTE`) — paiement commerçant déclenché après retrait confirmé.
      parameters: [{ $ref: "#/components/parameters/Id" }]
      requestBody: { $ref: "#/components/requestBodies/QrToken" }
      responses:
        "200": { description: Retrait confirmé }
        "409": { $ref: "#/components/responses/Error" }
  # ------------------------------------------------------------------ BANK
...
  /notifications:
    get:
      tags: [Notifications]
      summary: Mes notifications (polling)
      parameters:
        - { name: unread, in: query, schema: { type: string, enum: ["true", "false"] } }
        - { $ref: "#/components/parameters/Page" }
        - { $ref: "#/components/parameters/Limit" }
      responses: { "200": { description: Page de notifications } }
  /notifications/{id}/read:
    patch:
      tags: [Notifications]
      summary: Marque une notification comme lue
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses: { "200": { description: Lue }, "404": { $ref: "#/components/responses/Error" } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: string } }
    Page: { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
    Limit: { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
  requestBodies:
...
    QrToken:
      required: true
      content: { application/json: { schema: { type: object, required: [qrToken], properties: { qrToken: { type: string } } } } }
    Store:
      required: true
      content:
        application/json:
          schema:
            type: object
            required: [name, cityId, communeId]
            properties:
              name: { type: string }
              cityId: { type: string }
              communeId: { type: string }
              addressDescription: { type: string }
  responses:
    Error:
      description: "Erreur — `{ code, message, details?, requestId }`"
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  schemas:
...
    Error:
      type: object
      required: [code, message, requestId]
      properties:
        code: { type: string, example: VALIDATION_ERROR }
        message: { type: string }
        details: {}
        requestId: { type: string }
    OrderStatus:
      type: string
      enum: [CREEE, FINANCEMENT_EN_COURS, FINANCEE, PRETE_A_LIVRER, LIVREE, TERMINEE]
...